refactor web api test
